Finished dealer entry and dealer list.
+ Finished dealer creation page. e-mail address and split entry on this
page are both optional, the former entering "None Provided" when left
empty, and the latter searching the types table for a split value and
defaulting to .5 if no matching type is found.
+ + This means that the architecture of the program is moved away from the
sales report and toward the database. I had originally intended the split
to be the last thing found and calculated at the sales entry page for the
sake of calculating dealer accounts payable, but I've opted to form a
database that fills itself as you go along, not as you need it.
+ The dealer list page will list all dealers by ID number. I may create an
alternate list page that sorts by last name, as last name searching may be
used more frequently. A consignor may not remember his or her number, but
will definitely be able to provide all other data in the database from
memory.