Created sales widget / Standardized report dir
+ Decided to bring the naming scheme of the report directory in line with
that of the table directory, so the major division is singular (report,
table, application), and the subdirectories match the plurality of their
functions (table/dealers, table/sales, table/types, report/sales).
+ In terms of naming schemes, the view pages will match the subdirectory
name (dealers/dealer_update, etc.) and the controllers will match the
purpose of the view (dealers/update_controller, etc.). The widgets will
match the subdirectory name (sales/sales_widget, etc.).
+ Created the sales widget. The model may not be the most appropriate
place for the logic to unify the dealer_initials, _id, and _class, but
given that all pages in the sales directory will need that, I put it in
the model, rather than have a directory level controller that has just
that one function. I may change my mind a bit later down the road,
especially given the reports will need that function too, and thus the
application model might be the best place.