Added sales creation page
Man, was that a pain. Apparently DateTime doesn't help much in converting
a _POST string to an SQL date, so I'll have to go back and fix that in the
sales report generator. I'm planning on redoing the sales report generator
anyway, so no big deal.

+ Created date_fix in the application model, so I can call it to change a
mm/dd/YYYY to YYYY-mm-dd. Probably one of the more important things I've
done.

-Removed useless code from application model.