Changed date formatting on dealer report
+ I suppose I'm satisfied with how the dealer report works over all. I
really wanted to just change how the month is formatted for the query,
since I figured out the method I first implemented doesn't handle all
entry types well.
+ Deleted version.php. Anyone who wants a history and documentation of my
progress can just read the commit history.
+ Implemented a month and day fix to search month and days worth of data
in a query using the LIKE clause. The day fix is unused at this point in
time, but in the future a report may need it.